chore: update github actions to build flutter app instead of rust egui for macos/windows

+++ ui/flutter_app/lib/better_player_page.dart
@@ -682,7 +682,7 @@ class _BetterPlayerPageState extends State<BetterPlayerPage> with PlayerGestures
                   final delta = details.delta.dx;
                   final msDelta = (delta / totalWidth) * maxMs;
                   setState(() {
-                    _dragProgressPosition = (_dragProgressPosition + msDelta).clamp(0.0, maxMs);
+                    _dragProgressPosition = (_dragProgressPosition + msDelta).clamp(0.0, maxMs).toDouble();
                   });
                 },
                 onHorizontalDragEnd: (details) {
